Add `tcodec` package to handle timestamps during JSON decoding/encoding (#1202)
The way the extension works is by using a struct tag to define a `TimeCodec` to use for encoding/decoding a timestamp.
Normal `time.Time` fields can be used to define timestamps and the extension will handle proper JSON handling.
Configuration options allow to modify the extension behavior in detail to cover all use cases. 
 

This PR originates in ongoing work to support custom log parsers and can be used by both custom and 'normal' go log parsers. The code in this PR is isolated and does not affect the current log processor in any way.